+++ css/dashboard.css
@@ -55,7 +55,7 @@ footer {
     font-size: smaller;
 }
 
-@media screen and (max-width: 1400px) {
+@media screen and (max-width: 1210px) {
     .stTabs [data-baseweb="tab-list"] {
         padding-bottom: 25px;
     }
