+++ src/tabs/ai_readiness.py
@@ -1,5 +1,4 @@
 import streamlit as st
-from tabs.utils import create_html_table
 
 import pandas as pd
 import numpy as np
